fixed unconventional dots spacing in list of acronyms
Fixes the issue that an unconventional spacing is used between the dots in the list of acronyms. The dots are visible as soon as the `withpage` option is enabled. The applied fix changes the dot spacing to be the same as in all other lists, such as the table of contents or the list of listings.