enhance list of acronyms with hyperrefs
From the list of acronyms there are no links back to the places where the acronym appears for the first time. However, a user might expect this, since they are used to it from the other lists, e.g the table of content or the list of listings. This enhancement adds a similar functionality to the list of acronyms.